pyproject.toml
[tool.poetry]
name = "cv-cropper"
version = "0.1.0"
description = "Crops images using OpenCV-Python library and Streamlit"
authors = ["Luis Camargo <lcamargo@waveneuro.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.8.1.78"

[tool.poetry.group.dev.dependencies]
streamlit = "^1.28.2"
numpy = "^1.26.2"
Pillow = "^10.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
